refactor: application runner 통해 생성 책임 분리
